🗑️ cleanup chore
- Moved Router into s single directory
- Moved EmptyString into shared/data
- Moved dangling constants in shared/data to named directories
