--- libs/shared/data/src/lib/constants/application.constants.ts
@@ -1,4 +1,0 @@
-/**
- * Name of the application
- */
-export const ApplicationName = 'Itinerary Assistant';
