+++ libs/shared/data/src/lib/constants/application/application.constants.ts
@@ -1,0 +1,8 @@
+/**
+ * ### Name of the application
+ *
+ * ```ts
+ * const ApplicationName = 'Itinerary Assistant'
+ * ```
+ */
+export const ApplicationName = 'Itinerary Assistant';
