+++ libs/shared/data/src/lib/constants/index.ts
@@ -1,6 +1,11 @@
-export { ApplicationName } from './application.constants';
-export { Author } from './author.constants';
-export { FirebaseConfig } from './firebase.constants';
+export { ApplicationName } from './application/application.constants';
+
+export { Author } from './author/author.constants';
+
+export { FirebaseConfig } from './firebase/firebase.constants';
 
 export { TripListItemStub } from './trip/trip.constants';
+
 export { HttpStatus, ServerResponse } from './endpoint/endpoint.constants';
+
+export { EmptyString } from './string/string.constants';
