+++ libs/shared/data/src/lib/constants/string/string.constants.ts
@@ -1,0 +1,8 @@
+/**
+ * ### Empty String
+ *
+ * ```ts
+ * const EmptyString = ''
+ * ```
+ */
+export const EmptyString = '';
