+++ libs/ui/core/src/index.ts
@@ -1,5 +1,7 @@
 export { CoreModule } from './lib/core.module';
 
 export type { RouterPayloadModel } from './lib/models';
-export { EmptyString, RouterPayloadStub } from './lib/constants';
+
+export { RouterPayloadStub } from './lib/constants';
+
 export { CoreService, EndpointService, RouterService } from './lib/services';
