+++ libs/ui/core/src/lib/constants/index.ts
@@ -1,1 +1,1 @@
-export { RouterPayloadStub, EmptyString } from './router/router.constants';
+export { RouterPayloadStub } from '../services/router/constant/router.constants';
