+++ libs/ui/core/src/lib/imports/constants/index.ts
@@ -1,1 +1,1 @@
-export { HttpStatus } from '@itinerary-assistant/shared/data';
+export { EmptyString, HttpStatus } from '@itinerary-assistant/shared/data';
