+++ libs/ui/core/src/lib/services/router/index.ts
@@ -1,0 +1,5 @@
+export type { RouterPayloadModel } from './model/router.model';
+
+export { RouterPayloadStub } from './constant/router.constants';
+
+export { RouterService } from './service/router.service';
