+++ libs/ui/core/src/lib/services/router/model/router.model.ts
